+++ keybindings.el
@@ -485,7 +485,7 @@
  "9"	`(,(ilam (eepitch-ansi-term "9")) :which-key " ")
  "c"	'eab/switch-compile
  "x"	(ilam
-	 ;; (desktop-save (eab/desktop-dir))
+	 (desktop-save (eab/desktop-dir))
 	 (eab/sh-over-bash (concat "ssh chronos sudo systemctl restart " (cdr (assoc eab/daemon-name (gethash 'eab/emacs-service-alist eab/paths-hash)))) "" 't))
  "X"    (ilam
 	 (run-with-timer 0.1 nil 'kill-emacs)